POC: define and verify `boolAnd`
Draft, really needs to be cleaned up, especially needs better notations.
Allows to derive the semantics of a Sierra program by composing the
semantics of each instruction. Designed so that most of the work is done
by Lean itself, by unfolding equalities. My goal is to be able to
instrument that with metaprograms so that the semantics are easily
generated.